criado pastas e separando forma de lógica
