+++ Formulario/testando.py
@@ -1,30 +1,15 @@
 import json
 import os.path as os_path
-
-class Infrastructure:
-    def __init__(self, path_file, encoding):
-        self.path_file = path_file
-        self.encoding = encoding
-
-    def create_json_if_not_exist(self):
-        arquivo = os_path.isfile(f'./{self.path_file}')
-        create = self.path_file
-        if arquivo:
-            #pass
-            print('existe')
-        else:
-            with open(create, 'w') as arq:
-                dados = json.dumps([])
-                arq.write(dados)
-            print('arquivo criado')
-
+from infrastructure.testando2 import Infrastructure
 class Pessoas(Infrastructure):
-    def __init__(self,path_file, encoding):
-        super().__init__(path_file, encoding)
+    def __init__(self):
+        super().__init__()
         
     def funciona(self):
         print('funciona')
 
-arquivo = Pessoas('pessoas.json', encoding='utf-8')
-arquivo.create_json_if_not_exist()
-arquivo.funciona()
+
+if __name__ == '__main__':
+    arquivo = Pessoas()
+    arquivo.create_json_if_not_exist()
+    arquivo.funciona()
